Revert "Feature/syn 21 access point"
