--- .gitignore
@@ -1,2 +1,0 @@
-synestesia/libraries/*
-synestesia/vscode/*
