+++ synestesia/.gitignore
@@ -1,5 +1,2 @@
-.pio
-.vscode/.browse.c_cpp.db*
-.vscode/c_cpp_properties.json
-.vscode/launch.json
-.vscode/ipch
+libraries/*
+vscode/*
