--- synestesia/test/SYN-15-file-manager/data/global.syn
@@ -1,12 +1,0 @@
-# Global configuration for Synestesia
-# Configuration style is 
-#   Keyword : value (spaces are ignored)
-
-# SSID name
-SSID: Synestesia
-
-# PSWD for WiFi network. If NULL WiFi will be open
-PSWD: SYN2021
-
-# Sensorial module configuration path  (if any)
-SYNM: module/conf.syn
