--- synestesia/test/SYN-16-module-configuration/data/module.syn
@@ -1,9 +1,0 @@
-LIGHT:
-    CONN: 13,12,14
-    1:255,255,255
-    2:254,250,244
-
-LIGHT:
-    CONN: 13,12,15
-    1:255,255,255
-    2:0,123,123
