--- synestesia/test/SYN-21-access-point/data/global.syn
@@ -1,5 +1,0 @@
-# SSID name
-SSID: Synestesia
-
-# PSWD for WiFi network. If NULL WiFi will be open
-PSWD: SYN_2021
