--- synestesia/test/SYN-6-global-config/data/global.syn
@@ -1,4 +1,0 @@
-# SSID name
-SSID:Synestesia_2021
-# PSWD for WiFi network. If NULL WiFi will be open
-PSWD:SYN2021
