+++ synestesia/zip_libraries.sh
@@ -1,39 +1,22 @@
 if [ -d "libraries" ]
 then
-	echo "Removing existing libraries ..."
+	echo "Removing existing libraries ...\n"
 	rm -rf libraries/*
 else
-	echo "Directory libraries not found. Creating it ..."
+	echo "Directory libraries not found. Creating it ...\n"
 	mkdir libraries
 fi
 
 if ! [ -x "$(command -v zip)" ]
 then
-	echo "Could not found zip command. Installing it ..."
+	echo "Could not found zip command. Installing it ...\n"
 	sudo apt-get install zip
 fi
 
-echo "Zip main libraries"
 for lib in main/*
 do
-	if [ ! -f $lib ]
-	then
-		echo "* Zip $(basename $lib)"
-		zip -j libraries/"$(basename $lib).zip" $lib/*
-	fi
+	echo "$(basename $lib)" 
+	zip -j libraries/"$(basename $lib).zip" $lib/*
 done
 
-if [ -d "module" ]
-then
-	echo "Zip module libraries"
-	for lib in module/*
-	do
-		if [ ! -f "$lib" ]
-		then
-			echo "* Zip $(basename $lib)" 
-			zip -j libraries/"$(basename $lib).zip" $lib/*
-		fi
-	done
-fi
-
-echo "Libraries have been updated!"
+echo "\nLibraries have been updated!\n"
